@@ -1,5 +1,9 @@
-name: gdg_flutter_firebase_chat
+name: devfest_flutter_firebase_chat
 description: A new Flutter project.
+
+# The following line prevents the package from being accidentally published to
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -14,50 +18,56 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
+# Dependencies specify other packages that your package needs in order to work.
+# To automatically upgrade your package dependencies to the latest versions
+# consider running `flutter pub upgrade --major-versions`. Alternatively,
+# dependencies can be manually updated by changing the version numbers below to
+# the latest version available on pub.dev. To see which dependencies have newer
+# versions available, run `flutter pub outdated`.
 dependencies:
+  cached_network_image: ^3.1.0
+  cloud_firestore: ^2.5.4
+  cupertino_icons: ^1.0.2
+  firebase_auth: ^3.1.4
+  firebase_core: ^1.8.0
+  firebase_messaging: ^10.0.9
+  firebase_storage: ^10.0.6
   flutter:
     sdk: flutter
-
-  cloud_firestore: ^0.13.1
-  firebase_auth: ^0.15.4
-  firebase_core: ^0.4.4+3
-  firebase_messaging: ^6.0.9
-  firebase_storage: ^3.1.1
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-
+  flutter_image_compress: ^1.1.0
+  image_picker: ^0.8.4+3
+  intl: ^0.17.0
+  path_provider: ^2.0.6
+  provider: ^6.0.1
+  uuid: ^3.0.5
+ 
+flutter_localizations:
+  sdk: flutter
+  
 dev_dependencies:
+  flutter_lints: ^1.0.0
   flutter_test:
     sdk: flutter
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
-
   # To add assets to your application, add an assets section, like this:
   assets:
    - assets/images/
-
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
-
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
