@@ -1,14 +1,19 @@
+// ignore_for_file: constant_identifier_names
+
 import 'package:flutter/material.dart';
 
 class AppConstants {
-  static const String APP_PRIMARY_COLOR = "#EB342E";
+  static const String APP_PRIMARY_COLOR = "#026be4";
   static const String APP_BACKGROUND_COLOR = "#F6F8F9";
   static const String APP_BACKGROUND_COLOR_WHITE = "#FFFFFF";
   static const String APP_PRIMARY_COLOR_LIGHT = "#9f9f9f";
+
   static const String APP_PRIMARY_COLOR_BLACK = "#000000";
   static const String APP_PRIMARY_FONT_COLOR_WHITE = "#FFFFFF";
-  static const String APP_PRIMARY_COLOR_ACTION = "#BC2923";
-  static const String APP_PRIMARY_ROOM_COLOR = "#707070";
+  static const String APP_PRIMARY_FONT_COLOR_LIGHT = "#9f9f9f";
+
+  static const String APP_PRIMARY_COLOR_ACTION = "#026be4";
+  static const String APP_PRIMARY_TILE_COLOR = "#7eb1fb";
   static const String APP_PRIMARY_COLOR_GREEN = "#009099";
   static const String APP_BACKGROUND_COLOR_GRAY = "#D0D0D0";
 
