@@ -1,10 +1,88 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:gdg_flutter_firebase_chat/helpers/app_constants.dart';
-import 'package:gdg_flutter_firebase_chat/screens/chat_screen.dart';
+// import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:devfest_flutter_firebase_chat/helpers/app_constants.dart';
+import 'package:devfest_flutter_firebase_chat/screens/attendees_screen.dart';
+// import 'package:devfest_flutter_firebase_chat/screens/chat_screen.dart';
+import 'package:devfest_flutter_firebase_chat/screens/login_screen.dart';
+import 'package:devfest_flutter_firebase_chat/services/auth_service.dart';
+import 'package:devfest_flutter_firebase_chat/services/database_service.dart';
+import 'package:devfest_flutter_firebase_chat/services/storage_service.dart';
+import 'package:provider/provider.dart';
 
-void main() {
+import 'models/user_data.dart';
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
   runApp(
-    MaterialApp(
+    MultiProvider(
+      providers: [
+        ChangeNotifierProvider(
+          create: (_) => UserData(),
+        ),
+        Provider<AuthService>(
+          create: (_) => AuthService(),
+        ),
+        Provider<DatabaseService>(
+          create: (_) => DatabaseService(),
+        ),
+        Provider<StorageService>(
+          create: (_) => StorageService(),
+        ),
+      ],
+      child: const MyApp(),
+    ),
+  );
+}
+
+class MyApp extends StatefulWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  // final i18n = I18n.delegate;
+
+  @override
+  void initState() {
+    super.initState();
+    //  I18n.onLocaleChanged = onLocaleChange;
+  }
+
+  void onLocaleChange(Locale locale) {
+    setState(() {
+      // I18n.locale = locale;
+    });
+  }
+
+  Widget _getScreenId() {
+    return StreamBuilder<User?>(
+      stream: FirebaseAuth.instance.authStateChanges(),
+      builder: (BuildContext context, snapshot) {
+        if (snapshot.hasData) {
+          Provider.of<UserData>(context).currentUserId = snapshot.data!.uid;
+          return const AttendeesScreen();
+        } else {
+          return LoginScreen();
+        }
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      // localizationsDelegates: [
+      //   i18n,
+      //   GlobalMaterialLocalizations.delegate,
+      //   GlobalWidgetsLocalizations.delegate,
+      //   GlobalCupertinoLocalizations.delegate // <-- needed for iOS
+      // ],
+      // supportedLocales: i18n.supportedLocales,
       title: "GDG Firebase chat",
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
@@ -13,57 +91,24 @@ void main() {
             AppConstants.hexToColor(AppConstants.APP_BACKGROUND_COLOR),
         primaryColorLight:
             AppConstants.hexToColor(AppConstants.APP_PRIMARY_COLOR_LIGHT),
-        accentColor: Colors.black,
-        accentIconTheme: IconThemeData(color: Colors.black),
-        dividerColor: Colors.black12,
+        accentColor:
+            AppConstants.hexToColor(AppConstants.APP_PRIMARY_COLOR_BLACK),
+        accentIconTheme: IconThemeData(
+            color:
+                AppConstants.hexToColor(AppConstants.APP_PRIMARY_COLOR_BLACK)),
+        dividerColor:
+            AppConstants.hexToColor(AppConstants.APP_BACKGROUND_COLOR_GRAY),
         textTheme: TextTheme(
-          caption: TextStyle(color: Colors.white),
+          caption: TextStyle(
+              color: AppConstants.hexToColor(
+                  AppConstants.APP_PRIMARY_FONT_COLOR_WHITE)),
         ),
       ),
-      home: ChatScreen(),
-    ),
-  );
-}
-
-_appDrawer() {
-  return Drawer(
-    child: Column(
-      children: <Widget>[
-        DrawerHeader(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            children: <Widget>[
-              CircleAvatar(
-                radius: 30.0,
-                backgroundImage:
-                    AssetImage('assets/images/user_placeholder.jpg'),
-                backgroundColor: Colors.transparent,
-              ),
-              Text(
-                'Sumith Damodaran',
-                style: TextStyle(color: Colors.black),
-              ),
-              Text(
-                'PM @ Sitecore',
-                style: TextStyle(color: Colors.black),
-              )
-            ],
-          ),
-        ),
-        Spacer(),
-        ListTile(
-          leading: Icon(Icons.home),
-          title: Text('Home'),
-          onTap: () {},
-        ),
-        Divider(),
-        ListTile(
-          leading: Icon(Icons.people),
-          title: Text('Attendants'),
-          onTap: () {},
-        ),
-        Spacer(flex: 8),
-      ],
-    ),
-  );
+      home: _getScreenId(),
+      routes: {
+        LoginScreen.id: (context) => const LoginScreen(),
+        AttendeesScreen.id: (context) => const AttendeesScreen(),
+      },
+    );
+  }
 }
